Renamed all dark mode compatible elements.
