+++ UndertaleModTool/Editors/UndertaleEmbeddedAudioEditor.xaml
@@ -39,11 +39,11 @@
                 </TextBlock.Text>
             </TextBlock>
 
-            <local:Button Grid.Row="1" Grid.Column="0" Content="Import" Click="Import_Click"/>
-            <local:Button Grid.Row="1" Grid.Column="3" Content="Export" Click="Export_Click"/>
+            <local:ButtonDark Grid.Row="1" Grid.Column="0" Content="Import" Click="Import_Click"/>
+            <local:ButtonDark Grid.Row="1" Grid.Column="3" Content="Export" Click="Export_Click"/>
 
-            <local:Button Grid.Row="3" Grid.Column="0" x:Name="Play" Click="Play_Click">Play</local:Button>
-            <local:Button Grid.Row="3" Grid.Column="2" x:Name="Stop" Click="Stop_Click">Stop</local:Button>
+            <local:ButtonDark Grid.Row="3" Grid.Column="0" x:Name="Play" Click="Play_Click">Play</local:ButtonDark>
+            <local:ButtonDark Grid.Row="3" Grid.Column="2" x:Name="Stop" Click="Stop_Click">Stop</local:ButtonDark>
         </Grid>
     </Grid>
 </local:DataUserControl>
