+++ UndertaleModTool/Editors/UndertaleEmbeddedTextureEditor.xaml
@@ -79,8 +79,8 @@
                 </Border>
             </Viewbox>
 
-            <local:Button Grid.Row="1" Grid.Column="0" Content="Import" Click="Import_Click"/>
-            <local:Button Grid.Row="1" Grid.Column="3" Content="Export" Click="Export_Click"/>
+            <local:ButtonDark Grid.Row="1" Grid.Column="0" Content="Import" Click="Import_Click"/>
+            <local:ButtonDark Grid.Row="1" Grid.Column="3" Content="Export" Click="Export_Click"/>
         </Grid>
     </Grid>
 </local:DataUserControl>
