+++ UndertaleModTool/Editors/UndertaleExtensionEditor.xaml
@@ -37,7 +37,7 @@
         <local:UndertaleStringReference Grid.Row="2" Grid.Column="1" Margin="3" ObjectReference="{Binding ClassName}"/>
 
         <TextBlock Grid.Row="3" Grid.Column="0" Margin="3">Files</TextBlock>
-        <local:DataGrid Grid.Row="3" Grid.Column="1" Margin="3" ItemsSource="{Binding Files}" AutoGenerateColumns="False" CanUserAddRows="False" CanUserDeleteRows="True" HorizontalGridLinesBrush="LightGray" VerticalGridLinesBrush="LightGray" HeadersVisibility="Column" SelectionMode="Single" SelectionUnit="FullRow">
+        <local:DataGridDark Grid.Row="3" Grid.Column="1" Margin="3" ItemsSource="{Binding Files}" AutoGenerateColumns="False" CanUserAddRows="False" CanUserDeleteRows="True" HorizontalGridLinesBrush="LightGray" VerticalGridLinesBrush="LightGray" HeadersVisibility="Column" SelectionMode="Single" SelectionUnit="FullRow">
             <DataGrid.Columns>
                 <DataGridTemplateColumn Header="Files" Width="*">
                     <DataGridTemplateColumn.CellTemplate>
@@ -47,10 +47,10 @@
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
             </DataGrid.Columns>
-        </local:DataGrid>
+        </local:DataGridDark>
 
         <TextBlock Grid.Row="4" Grid.Column="1" Margin="5" TextAlignment="Center" FontStyle="Italic">Click button below to add extension file.</TextBlock>
-        <local:Button Grid.Row="5" Grid.Column="1" Content="Add new extension file" Click="Button_Click"/>
+        <local:ButtonDark Grid.Row="5" Grid.Column="1" Content="Add new extension file" Click="Button_Click"/>
 
         <TextBlock Grid.Row="6" Grid.Column="0" Margin="3" Visibility="{Binding DataContext.IsExtProductIDEligible, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:MainWindow}}}">Product ID</TextBlock>
         <TextBox Background="#FF333333" Foreground="#FFFFFFFF" Grid.Row="6" Grid.Column="1" Margin="3" Visibility="{Binding DataContext.IsExtProductIDEligible, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:MainWindow}}}" Text="{Binding ProductIdData, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:UndertaleExtensionEditor}}, Mode = TwoWay, Converter={StaticResource byteArrayConverter}}"/>
