+++ UndertaleModTool/Editors/UndertaleSpriteEditor.xaml
@@ -91,8 +91,8 @@
         <TextBox Grid.Row="7" Grid.Column="1" Margin="3" Text="{Binding BBoxMode}"/>
 
         <TextBlock Grid.Row="8" Grid.Column="0" Margin="3">SepMasks</TextBlock>
-        <local:ComboBox ItemsSource="{Binding Source={StaticResource SepMaskType}}" SelectedItem="{Binding Path=SepMasks}"
-                        Grid.Row="8" Grid.Column="1" Margin="3"/>
+        <local:ComboBoxDark ItemsSource="{Binding Source={StaticResource SepMaskType}}" SelectedItem="{Binding Path=SepMasks}"
+                            Grid.Row="8" Grid.Column="1" Margin="3"/>
 
         <TextBlock Grid.Row="9" Grid.Column="0" Margin="3">Origin</TextBlock>
         <Grid Grid.Row="9" Grid.Column="1">
@@ -105,10 +105,10 @@
         </Grid>
 
         <TextBlock Grid.Row="10" Grid.Column="0" Margin="3">Textures</TextBlock>
-        <local:DataGrid Grid.Row="10" Grid.Column="1" Margin="3" ItemsSource="{Binding Textures, Mode=OneWay}" x:Name="TextureList"
-                        AutoGenerateColumns="False" CanUserAddRows="True" CanUserDeleteRows="True"
-                        HorizontalGridLinesBrush="LightGray" VerticalGridLinesBrush="LightGray"
-                        HeadersVisibility="None" SelectionMode="Single" SelectionUnit="FullRow" IsSynchronizedWithCurrentItem="True">
+        <local:DataGridDark Grid.Row="10" Grid.Column="1" Margin="3" ItemsSource="{Binding Textures, Mode=OneWay}" x:Name="TextureList"
+                            AutoGenerateColumns="False" CanUserAddRows="True" CanUserDeleteRows="True"
+                            HorizontalGridLinesBrush="LightGray" VerticalGridLinesBrush="LightGray"
+                            HeadersVisibility="None" SelectionMode="Single" SelectionUnit="FullRow" IsSynchronizedWithCurrentItem="True">
             <DataGrid.Resources>
                 <SolidColorBrush x:Key="{x:Static SystemColors.HighlightBrushKey}" Color="#FF26A0DA"/>
                 <Style TargetType="{x:Type DataGridCell}">
@@ -146,7 +146,7 @@
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
             </DataGrid.Columns>
-        </local:DataGrid>
+        </local:DataGridDark>
 
         <Viewbox Grid.Row="11" Grid.Column="1" Stretch="Uniform" StretchDirection="DownOnly">
             <Border>
@@ -164,10 +164,10 @@
             </Border>
         </Viewbox>
 
-        <local:Button Grid.Row="12" Grid.Column="1" Content="Export all frames (or Spine data if this is a Spine sprite)" Click="ExportAll_Click"/>
+        <local:ButtonDark Grid.Row="12" Grid.Column="1" Content="Export all frames (or Spine data if this is a Spine sprite)" Click="ExportAll_Click"/>
 
         <TextBlock Grid.Row="13" Grid.Column="0" Margin="3">Collision masks</TextBlock>
-        <local:DataGrid Grid.Row="13" Grid.Column="1" Margin="3" ItemsSource="{Binding CollisionMasks, Mode=OneWay}" x:Name="MaskList" AutoGenerateColumns="False" CanUserAddRows="True" CanUserDeleteRows="True" HorizontalGridLinesBrush="LightGray" VerticalGridLinesBrush="LightGray" HeadersVisibility="None" SelectionMode="Single" SelectionUnit="FullRow" IsSynchronizedWithCurrentItem="True" AddingNewItem="MaskList_AddingNewItem">
+        <local:DataGridDark Grid.Row="13" Grid.Column="1" Margin="3" ItemsSource="{Binding CollisionMasks, Mode=OneWay}" x:Name="MaskList" AutoGenerateColumns="False" CanUserAddRows="True" CanUserDeleteRows="True" HorizontalGridLinesBrush="LightGray" VerticalGridLinesBrush="LightGray" HeadersVisibility="None" SelectionMode="Single" SelectionUnit="FullRow" IsSynchronizedWithCurrentItem="True" AddingNewItem="MaskList_AddingNewItem">
             <DataGrid.Resources>
                 <SolidColorBrush x:Key="{x:Static SystemColors.HighlightBrushKey}" Color="#FF26A0DA"/>
                 <Style TargetType="{x:Type DataGridCell}">
@@ -202,7 +202,7 @@
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
             </DataGrid.Columns>
-        </local:DataGrid>
+        </local:DataGridDark>
 
         <ContentControl Grid.Row="14" Grid.Column="1" Margin="3" DataContext="{Binding SelectedItem, Mode=OneWay, ElementName=MaskList}">
             <ContentControl.Template>
@@ -238,8 +238,8 @@
                                 <RowDefinition Height="Auto"/>
                             </Grid.RowDefinitions>
 
-                            <local:Button Grid.Row="0" Grid.Column="0" Content="Import mask" Click="MaskImport_Click" x:Name="MaskImport"/>
-                            <local:Button Grid.Row="0" Grid.Column="3" Content="Export mask" Click="MaskExport_Click" x:Name="MaskExport"/>
+                            <local:ButtonDark Grid.Row="0" Grid.Column="0" Content="Import mask" Click="MaskImport_Click" x:Name="MaskImport"/>
+                            <local:ButtonDark Grid.Row="0" Grid.Column="3" Content="Export mask" Click="MaskExport_Click" x:Name="MaskExport"/>
                         </Grid>
                     </StackPanel>
                     <ControlTemplate.Triggers>
@@ -283,8 +283,8 @@
             <TextBox Grid.Row="0" Grid.Column="1" Margin="3" Text="{Binding SVersion}"/>
 
             <TextBlock Grid.Row="1" Grid.Column="0" Margin="3">Type</TextBlock>
-            <local:ComboBox ItemsSource="{Binding Source={StaticResource SpriteType}}" SelectedItem="{Binding Path=SSpriteType}"
-                            Grid.Row="1" Grid.Column="1" Margin="3"/>
+            <local:ComboBoxDark ItemsSource="{Binding Source={StaticResource SpriteType}}" SelectedItem="{Binding Path=SSpriteType}"
+                                Grid.Row="1" Grid.Column="1" Margin="3"/>
 
             <Grid Grid.Row="2" Grid.ColumnSpan="2" Margin="0" Visibility="{Binding DataContext.IsGMS2, Mode=OneTime, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type local:MainWindow}}}">
                 <Grid.ColumnDefinitions>
@@ -300,8 +300,8 @@
                 <TextBox Grid.Row="0" Grid.Column="1" Margin="3" Text="{Binding GMS2PlaybackSpeed}"/>
 
                 <TextBlock Grid.Row="1" Grid.Column="0" Margin="3">Playback speed type</TextBlock>
-                <local:ComboBox ItemsSource="{Binding Source={StaticResource AnimSpeedType}}" SelectedItem="{Binding Path=GMS2PlaybackSpeedType}"
-                                Grid.Row="1" Grid.Column="1" Margin="3"/>
+                <local:ComboBoxDark ItemsSource="{Binding Source={StaticResource AnimSpeedType}}" SelectedItem="{Binding Path=GMS2PlaybackSpeedType}"
+                                    Grid.Row="1" Grid.Column="1" Margin="3"/>
             </Grid>
         </Grid>
     </Grid>
