+++ UndertaleModTool/Editors/UndertaleTexturePageItemEditor.xaml
@@ -87,8 +87,8 @@
                 <ColumnDefinition Width="*"/>
             </Grid.ColumnDefinitions>
 
-            <local:Button Grid.Column="0" Content="Import" Click="Import_Click"/>
-            <local:Button Grid.Column="3" Content="Export" Click="Export_Click"/>
+            <local:ButtonDark Grid.Column="0" Content="Import" Click="Import_Click"/>
+            <local:ButtonDark Grid.Column="3" Content="Export" Click="Export_Click"/>
         </Grid>
 
         <Viewbox Grid.Row="6" Grid.Column="0" Grid.ColumnSpan="2" Stretch="Uniform" StretchDirection="DownOnly"
