+++ UndertaleModTool/Editors/UndertaleVariableEditor.xaml
@@ -32,7 +32,7 @@
         <local:UndertaleStringReference Grid.Row="0" Grid.Column="1" Margin="3" ObjectReference="{Binding Name}"/>
 
         <TextBlock Grid.Row="1" Grid.Column="0" Margin="3">Instance type</TextBlock>
-        <local:ComboBox Grid.Row="1" Grid.Column="1" Margin="3" ItemsSource="{Binding Source={StaticResource InstanceType}}" SelectedItem="{Binding InstanceType}" />
+        <local:ComboBoxDark Grid.Row="1" Grid.Column="1" Margin="3" ItemsSource="{Binding Source={StaticResource InstanceType}}" SelectedItem="{Binding InstanceType}" />
 
         <TextBlock Grid.Row="2" Grid.Column="0" Margin="3">VarID</TextBlock>
         <TextBox Grid.Row="2" Grid.Column="1" Margin="3" Text="{Binding VarID}"/>
