+++ UndertaleModTool/Windows/StringUpdateWindow.xaml
@@ -15,8 +15,8 @@
             <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
         <TextBlock Grid.Row="0" Margin="3">You edited a string reference! What do you want to do?</TextBlock>
-        <local:Button Grid.Row="1" Margin="3" Click="Button1_Click">Change only this one value</local:Button>
-        <local:Button Grid.Row="2" Margin="3" Click="Button2_Click">Change all occurrences of this value</local:Button>
-        <local:Button Grid.Row="3" Margin="3" Click="Button3_Click">Abort the change</local:Button>
+        <local:ButtonDark Grid.Row="1" Margin="3" Click="Button1_Click">Change only this one value</local:ButtonDark>
+        <local:ButtonDark Grid.Row="2" Margin="3" Click="Button2_Click">Change all occurrences of this value</local:ButtonDark>
+        <local:ButtonDark Grid.Row="3" Margin="3" Click="Button3_Click">Abort the change</local:ButtonDark>
     </Grid>
 </Window>
